TINY-3909: Move remove_trailing_brs node filter from DomParser to DomSerializer
